fix(nfc): diagnostico y workarounds HCE para PN532 v1.6 + ajustes app/backend
- Cambia libreria PN532 de Adafruit a elechouse/Seeed para mejor soporte ISO-DEP.
- Aumenta timeout de readPassiveTargetID a 500ms y mejora diagnostico de inDataExchange.
- Anade Le=0x00 al SELECT AID para compatibilidad APDU case 4.
- HceService: vibracion + Toast diagnostico al recibir APDU.
- MainActivity: vibracion/Toast al activar setPreferredService (exito vs fallo).
- AndroidManifest: permiso VIBRATE.
- Mueve cod.ino a codigo_sensores/cod/cod.ino.
- Anade **/build/ al .gitignore.
- Cambios menores en app Flutter, modelos y rutas backend.

+++ aula_inteligente/lib/features/environment/models/sensor_data_model.dart
@@ -42,7 +42,7 @@ class SensorData {
     return SensorData(
       temperature: (map['temperature'] ?? 0.0).toDouble(),
       humidity: (map['humidity'] ?? 0.0).toDouble(),
-      airQualityIndex: (map['air_quality_index'] ?? 0).toInt(),
+      airQualityIndex: (map['air_quality_index'] ?? 0.0).toDouble(),
       smokeDetected: map['smoke_detected'] ?? false,
       powerConsumptionWatts: (map['power_consumption_watts'] ?? 0).toDouble(),
       timestamp: DateTime.tryParse(map['timestamp'] ?? '') ?? DateTime.now(),
